==> Scrapy/04. BeautifulSoup로 스크레이핑하기.ipynb <==
### BeautifulSoup
- 간단하게 HTML과 XML에서 정보를 추출

# !pip install beautifulsoup4
# !pip install lxml

from bs4 import BeautifulSoup
import urllib.request as req

url = "https://zeushahn.github.io/Test/python/bs_exam01.html"

# res = 소스 가져오기
res = req.urlopen(url)
print(res)

# BeautifulSoup으로 분석하기
soup = BeautifulSoup(res,"html.parser")
print(soup)

# 원하는 부분 추출하기 (스크레이핑이란?)
h1 = soup.html.body.h1
# 요소의 글자 출력하기 : String, text
print("h1 =",h1.string)

# 원하는 부분 추출하기 (웹 페이지를 분석하는 것)
p1 = soup.html.body.p
# 요소의 글자 출력하기 : String, text
print("p1 =",p1.text)

# 다음번째 p 가져올 때:
# </p> (닫는 태그도 카운트하기 때문에 next_sibling 두 번 사용한다.)
p2 = p1.next_sibling.next_sibling
# 요소의 글자 출력하기 : String, text
print("p2 =",p2.text)

# id가 있을 때 id로 해당 부분 추출하기
from bs4 import BeautifulSoup
import urllib.request as req
url = "https://zeushahn.github.io/Test/python/bs_exam02.html"

# res = 소스 가져오기
res = req.urlopen(url)

# BeautifulSoup으로 분석하기
soup = BeautifulSoup(res,"html.parser")

title = soup.find(id='title')
print(title.text)

from bs4 import BeautifulSoup
import urllib.request as req
url = "https://zeushahn.github.io/Test/python/bs_exam03.html"

# res = 소스 가져오기
res = req.urlopen(url)


# BeautifulSoup으로 분석하기
soup = BeautifulSoup(res,"html.parser")


links = soup.find_all("a")
for a in links:
    text = a.string
    href = a.attrs['href']
    print(text,">",href)


from bs4 import BeautifulSoup
import urllib.request as req
url = "https://www.kma.go.kr/weather/forecast/mid-term-rss3.jsp?stnId=109"

# res = 소스 가져오기
res = req.urlopen(url)

# BeautifulSoup으로 분석하기
soup = BeautifulSoup(res,"html.parser")

links = soup.find("title")
print(links.string)


from bs4 import BeautifulSoup
import urllib.request as req
url = "https://www.kma.go.kr/weather/forecast/mid-term-rss3.jsp?stnId=109"

# res = 소스 가져오기
res = req.urlopen(url)

# BeautifulSoup으로 분석하기
soup = BeautifulSoup(res,"html.parser")

links = soup.rss.channel.item.description.header.wf.text
for wf in links:
    links = links.replace("○","\n")
    links = links.replace("<br />","")
    links = links.replace("          ","")
    
    
    

print(links)


### CSS 선택자 사용하기
- BeautifulSoup은 JavaScript Library인 JQuery처럼 CSS 선택자를 지정해서 원하는 요소를 추출 가능.

from bs4 import BeautifulSoup
import urllib.request as req
url = "https://zeushahn.github.io/Test/python/bs_exam04.html"

# res = 소스 가져오기
res = req.urlopen(url)


# BeautifulSoup으로 분석하기
soup = BeautifulSoup(res,"html.parser")

print(soup)

### 필요한 부분을 CSS Query로 추출하기 (#: id, .: class, >: 자식, 빈칸: 후손)

# id가 meigen인 div의 하위에 있는 (자식인) h1
h1 = soup.select_one("div#meigen > h1").string
print("h1 =",h1)

# id가 meigen인 div의 하위에 있는 ul 클래스 중 id가 items인 ul을 찾고 그 하위의 li를 전부 출력
li_list = soup.select("div#meigen > ul.items > li")
for li in li_list:
    print("li =",li.string)

#exchangeList > li.on > a.head.usd > div > span.value

from bs4 import BeautifulSoup
import urllib.request as req
url = "https://finance.naver.com/marketindex/"

# res = 소스 가져오기
res = req.urlopen(url)


# BeautifulSoup으로 분석하기
soup = BeautifulSoup(res,"html.parser")

price = soup.select_one("#exchangeList > li.on > a.head.usd > div > span.value").string
print("usd/krw",price)

# 미국, 일본, 유럽연합, 중국 환율
#exchangeList > li.on > a.head.usd > div > span.value
#exchangeList > li.on > a.head.jpy > div > span.value
#exchangeList > li.on > a.head.eur > div > span.value
#exchangeList > li.on > a.head.cny > div > span.value

from bs4 import BeautifulSoup
import urllib.request as req
url = "https://finance.naver.com/marketindex/"

# res = 소스 가져오기
res = req.urlopen(url)

# BeautifulSoup으로 분석하기
soup = BeautifulSoup(res,"html.parser")
exchange = soup.select(".value")
country = ['미국','일본','유럽연합','중국']
i = 0
for ex in exchange:
    print(country[i],'\t:',ex.string.rjust(10))
    i += 1
    if i >= len(country):
        break

---
### 다음영화 연간 순위 가져오기

#mainContent > div > div.box_boxoffice > ol > li:nth-child(1) > div > div.thumb_cont > strong > a
#mainContent > div > div.box_boxoffice > ol > li:nth-child(50) > div > div.thumb_cont > strong > a
#mainContent > div > div.box_boxoffice > ol > li:nth-child(1) > div > div.thumb_cont > strong > a
from bs4 import BeautifulSoup
import urllib.request as req
import time

# url = "https://movie.daum.net/ranking/boxoffice/yearly"
# url = "https://movie.daum.net/ranking/boxoffice/yearly?date=2021"

# 년도 받아오기
prompt = """
1. 올해 순위 보기
2. 예전 순위 보기
0. 나가기
"""
num = 10
while num != 0:
    print(prompt)
    num = int(input())
    if num == 1:
        url = "https://movie.daum.net/ranking/boxoffice/yearly"
        # res = 소스 가져오기
        res = req.urlopen(url)

        # BeautifulSoup으로 분석하기
        soup = BeautifulSoup(res,"html.parser")
        i = list(range(1,11))
        for i in i:
            time.sleep(0.1)
            rank = soup.select_one("#mainContent > div > div.box_boxoffice > ol > li:nth-child(%s) > div > div.thumb_cont > strong > a"%i)
            print(i,'위:\t',rank.text)
        continue
     
    elif num == 2:
        print('년도를 입력하세요')
        year = int(input())
        if year in range(2004,2022):
            url = "https://movie.daum.net/ranking/boxoffice/yearly?date=%d"%year

            # res = 소스 가져오기
            res = req.urlopen(url)

            # BeautifulSoup으로 분석하기
            soup = BeautifulSoup(res,"html.parser")
            i = list(range(1,11))
            for i in i:
                time.sleep(0.1)
                rank = soup.select_one("#mainContent > div > div.box_boxoffice > ol > li:nth-child(%s) > div > div.thumb_cont > strong > a"%i)
                print(i,'위:\t',rank.text)
            continue
        else:
            break
    
        
    elif num == 0:
        break

    else:
        print('번호를 확인해주세요')
        continue

import urllib.parse as parse
import urllib.request as req
# 윤동주 문서 가져오기
#mw-content-text > div.mw-parser-output > ul:nth-child(6) > li > b > a
# => mw-content-text > div.mw-parser-output > ul > li > b > a

# parse를 사용한 인코딩
# parse할 부분을 분리
import math
url = "https://ko.wikisource.org/wiki/"
auth = '저자'
name = '윤동주'
url = url + parse.quote(auth) + ":" + parse.quote(name)

res = req.urlopen(url)
soup = BeautifulSoup(res,"html.parser")
# poem = soup.select("#mw-content-text > div.mw-parser-output > ul > li > b > a")
# bold체만 가져옴.
# => li 하위 후손들을 가져오도록 수정
poem = soup.select("#mw-content-text > div.mw-parser-output > ul > li a")
for po in poem:
    name = po.text
    if name == "증보판": # 하늘과 바람과 별과 시 옆의 괄호내용임
        continue
    print("-",name)
        
   

# 다음영화 주간 순위
# 순위
#mainContent > div > div.box_boxoffice > ol > li:nth-child(4) > div > div.thumb_item > div.poster_movie > span.rank_num
# 제목
#mainContent > div > div.box_boxoffice > ol > li:nth-child(1) > div > div.thumb_cont > strong > a

url = "https://movie.daum.net/ranking/boxoffice/weekly"
# ?date=20230105

res = req.urlopen(url)
soup = BeautifulSoup(res,"html.parser")
r_list = soup.select("#mainContent > div > div.box_boxoffice > ol > li > div > div.thumb_item > div.poster_movie > span.rank_num")
t_list = soup.select("#mainContent > div > div.box_boxoffice > ol > li > div > div.thumb_cont > strong > a")
rank = [int(r.text) for r in r_list]
title = [t.text for t in t_list]
print(rank)
print(title)

# 다음영화 주간 순위
# 순위
#mainContent > div > div.box_boxoffice > ol > li:nth-child(4) > div > div.thumb_item > div.poster_movie > span.rank_num
# 제목
#mainContent > div > div.box_boxoffice > ol > li:nth-child(1) > div > div.thumb_cont > strong > a

url = "https://movie.daum.net/ranking/boxoffice/weekly"
# ?date=20230105

res = req.urlopen(url)
soup = BeautifulSoup(res,"html.parser")
# 클래스 이름으로 불러오는 방법도 있음
t_list = soup.select(".link_txt")[0:15]
titleList = [t.text for t in t_list]
print(titleList)

# countLists만으로 Dataframe 생성
import pandas as pd

# 다음영화 주간 순위
# 순위
#mainContent > div > div.box_boxoffice > ol > li:nth-child(4) > div > div.thumb_item > div.poster_movie > span.rank_num
# 제목
#mainContent > div > div.box_boxoffice > ol > li:nth-child(1) > div > div.thumb_cont > strong > a

url = "https://movie.daum.net/ranking/boxoffice/weekly"
# ?date=20230105

res = req.urlopen(url)
soup = BeautifulSoup(res,"html.parser")
r_list = soup.select("#mainContent > div > div.box_boxoffice > ol > li > div > div.thumb_item > div.poster_movie > span.rank_num")
t_list = soup.select("#mainContent > div > div.box_boxoffice > ol > li > div > div.thumb_cont > strong > a")
rank = [int(r.text) for r in r_list]
title = [t.text for t in t_list]
print(rank)
print(title)
df = pd.DataFrame(data=rank,columns=["Index"])
df


# df Dataframe에 컬럼 추가
df['Titles'] = pd.Series(title)
df

# zip을 사용해서 row 한 줄의 data를 지정한다.
df2 = pd.DataFrame(data=zip(rank,title), columns=["Index","Title"], index=None)
df2

# csv로 저장
df2.to_csv("../Data/DaumList.csv",encoding="utf-8",index=False)
df2

pd.read_csv("../Data/DaumList.csv",index_col=0)

# 네이버 영화 순위 불러오기 => df => csv 저장
import pandas as pd
from bs4 import BeautifulSoup
import urllib.request as req
#old_content > table > tbody > tr:nth-child(2) > td:nth-child(1) > img
#old_content > table > tbody > tr > td.title > div > a
url = "https://movie.naver.com/movie/sdb/rank/rmovie.naver"

res = req.urlopen(url)
soup = BeautifulSoup(res,"html.parser")

# 순위
ranks = soup.select("#old_content > table > tbody > tr > td > img")
r_list = [r.attrs['alt'] for r in ranks]

# 순위에서 숫자만 추출 (변동폭 제외)
# new_r_list = []
# for r in r_list:
#     try:
#         new_r_list.append(int(r))
#     except:
#         pass
# print(new_r_list)

# 순위에 빈 칸이 있으므로 숫자 직접 지정
new_r_list = list(range(1,51))
print(new_r_list)

# 제목
titles = soup.select("td.title")
t_list = [t.text.replace("\n","") for t in titles]
print(t_list)

# Dataframe
df = pd.DataFrame(data=zip(new_r_list,t_list),columns=["Index","Title"])

df.to_csv("../Data/NaverList.csv",index=None)
pd.read_csv("../Data/NaverList.csv",index_col=0)
