==> Scrapy/05. Selenium.ipynb <==
# !pip install selenium

# !pip install webdriver-manager

from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.by import By

---
### 오피넷 유가정보

# Chrome Browser과 Chrome Driver Version 확인하기
chrom_options = webdriver.ChromeOptions()
driver = webdriver.Chrome(service=Service(ChromeDriverManager().install()),options=chrom_options)
driver.get("https://www.opinet.co.kr/searRgSelect.do")

# 현재 화면 Capture하기
# 나에게 제어되고 있는 브라우저인지 확인
driver.save_screenshot("../Data/001.png")

---
# 서울시 유가 정보 획득을 위한 Test

### 자치구명 목록 가져오기

# 자치구명 정보 가져오기 (XPath): //*[@id="SIGUNGU_NM0"]
gu_list_raw = driver.find_element(By.XPATH, '//*[@id="SIGUNGU_NM0"]')
gu_list = gu_list_raw.find_elements(By.TAG_NAME,'option')
gu_names = [option.get_attribute('value') for option in gu_list]
gu_names.remove('')
print(gu_names)


### 조회버튼 클릭

# 조회버튼 click위해 XPath 가져오기: //*[@id="searRgSelect"]/span
xpath = '//*[@id="searRgSelect"]/span'

search = driver.find_element(By.XPATH, xpath).click()
# 실행하면 조회버튼 클릭 실행됨

print(gu_names)

### 자치구명 선택해서 전달

# 자치구 이름을 획득하여 1번 자치구를 실행
element = driver.find_element(By.ID, 'SIGUNGU_NM0')
# 드랍다운에 key값을 전해준다. (1번: 강동구)
element.send_keys(gu_names[1])

### 엑셀 저장버튼 클릭

# 엑셀 저장버튼 클릭: //*[@id="glopopd_excel"]/span
save_path = '//*[@id="glopopd_excel"]/span'
save = driver.find_element(By.XPATH, save_path).click()

# 테스트 완료

---

# 데이터 수집
- 서울시 1번 ~ 25번 자치구별 주유소 조회 후 엑셀파일로 저장까지

# 화면에 그림을 그리는 시간이 필요하므로 이를 delay 시키도록 한다.
# import time
# for gu in gu_names:
#     element = driver.find_element(By.ID, 'SIGUNGU_NM0')
#     element.send_keys(gu)
#     time.sleep(2)
    
#     xpath = '//*[@id="SIGUNGU_NM0"]'
#     gu_list_raw = driver.find_element(By.XPATH, xpath)
#     time.sleep(1)

#     xpath = '//*[@id="glopopd_excel"]/span'
#     element_get_excel = driver.find_element(By.XPATH, save_path).click()
#     time.sleep(1)

---
# 가장 맛있는 족발 매장정보
- 매장명
- 매장주소
- 전화번호
- 이미지


from selenium import webdriver
from selenium.webdriver.chrome.service import Service
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.by import By
import pandas as pd
from bs4 import BeautifulSoup
import urllib.request as req
import time

chrom_options = webdriver.ChromeOptions()
driver = webdriver.Chrome(service=Service(ChromeDriverManager().install()),options=chrom_options)
driver.get("https://gajok.kr/sub/store.php")

url = "https://gajok.kr/sub/store.php"
res = req.urlopen(url)
soup = BeautifulSoup(res,"html.parser")

names = []
addresses = []
tels = []
# 상세정보 버튼 클릭
for i in range(1,404):
    try:
        
        driver.find_element(By.XPATH,f'//*[@id="result_store"]/div[1]/ol/li[{i}]/a').click()
        time.sleep(1)
    
        # 정보 dialog
        info_list_raw = driver.find_element(By.XPATH, '//*[@id="map"]/div[1]/div/div[6]/div[404]')
    
        # 지점명 가져오기
        name_list = info_list_raw.find_element(By.TAG_NAME,'p').get_attribute('textContent')
        print('지점명:\t',name_list)
        names.append(name_list)

        # 주소 가져오기
        address_list = info_list_raw.find_element(By.TAG_NAME,'li').get_attribute('textContent')
        print('주소:\t',address_list)
        time.sleep(1)
        addresses.append(address_list)

        # tel_list = soup.find('#map > div > div > div > div > div > div.text > ul > li')
        tel_list_raw = driver.find_element(By.XPATH,'//*[@id="map"]/div[1]/div/div[6]/div[404]/div/div[2]/ul/li[2]')
        tel_list = tel_list_raw.get_attribute("textContent")
        print('전화번호:\t',tel_list)
        tels.append(tel_list)

        driver.find_element(By.XPATH,'//*[@id="title"]/a/img').click()
        time.sleep(1)
    except:
        print("----------\n오류\n----------")
        pass
    


df = pd.DataFrame(data=zip(names,addresses,tels), columns=["Branch","Address","Tel"], index=None)
df
df.to_csv("../Data/Gajok.csv",encoding="utf-8",index=False)
df
