==> Scrapy/07. Database.ipynb <==
### 데이터베이스

SQLite  
: 가볍게 파일 하나로 사용할 수 있는 DB

import sqlite3

# sqlite 데이터베이스 연결하기
dbPath = "../Data/test.sqlite"
conn = sqlite3.connect(dbPath)

# 테이블을 생성하고 데이터 넣기
curs = conn.cursor()
# executescript는 여러개 실행할 때
# 하나라면 execute
curs.executescript(
    """
    /* items 테이블이 이미 있다면 제거하기 */
    DROP TABLE IF EXISTS items;

    /* 테이블 생성하기 */
    CREATE TABLE items(
        item_id INTEGER PRIMARY KEY AUTOINCREMENT,
        name TEXT UNIQUE,
        price INTEGER
    );

    /* 데이터 넣기 */
    INSERT INTO items(name, price) VALUES ('Apple',800);
    INSERT INTO items(name, price) VALUES ('Orange',780);
    INSERT INTO items(name, price) VALUES ('Banana',430);
    """
)

# 위의 작업을 데이터베이스에 반영하기
conn.commit()
# commit 실행시 memory => 파일 저장



# 데이터 추출하기
curs = conn.cursor()
curs.execute("SELECT * FROM items")
item_list = curs.fetchall()
print(item_list)
# 결과: 튜플로 저장되있음
for i in item_list:
    print(i[0],i[1],i[2])

filePath = "../Data/test2.sqlite"
conn = sqlite3.connect(filePath)

# 테이블 생성하기
curs = conn.cursor()
curs.execute("DROP TABLE IF EXISTS items")
curs.execute(
    """
    CREATE TABLE items(
        item_id INTEGER PRIMARY KEY AUTOINCREMENT,
        name TEXT UNIQUE,
        price INTEGER
    );
    """
)
conn.commit()


# 데이터 넣기 (다른 곳에서 데이터 받아서 넣을 때)
curs = conn.cursor()
curs.execute("INSERT INTO items (name, price) VALUES (?,?)",("Orange",5200))
conn.commit()


curs = conn.cursor()
curs.execute("SELECT * FROM items")
item_list = curs.fetchall()
print(item_list)
# 결과: 튜플로 저장되있음
for i in item_list:
    print(i[0],i[1])

# 여러 데이터를 연속으로 넣기
curs = conn.cursor()
# 데이터부터 만듬 (여러 데이터)
data = [("Mango",7700),("Kiwi",4000),("Peach",9400),("Persimmon",7000),("Banana",4000)]
curs.executemany("INSERT INTO items(name,price) VALUES (?,?)", data)
conn.commit()


curs.execute("SELECT * FROM items")
item_list = curs.fetchall()
print(item_list)
# 결과: 튜플로 저장되있음
for i in item_list:
    print(i[0],i[1])

# 4000~7000원 사이의 데이터 추출하기
curs = conn.cursor()
curs.execute("SELECT * FROM items WHERE price >= ? and price <= ?",("4000","7000"))
selected_item_list = curs.fetchall()
print(selected_item_list)
# 결과: 튜플로 저장되있음
for i in selected_item_list:
    print(i[0],i[1])

---
### MySQL

# !pip install pymysql

---
### 프로젝트: Flutter-Python Server (RestAPI)-Database
---

import pymysql

# Connection
conn = pymysql.connect(
    host = "127.0.0.1",
    user = "root",
    password = "qwer1234",
    db = "education",
    charset = "utf8"
)

# Cursor
curs = conn.cursor()


sql = "select * from student"
curs.execute(sql)
rows = curs.fetchall()
row_list = list(rows)
print(row_list)

print(row_list[0])
print(row_list[0][1])

# Insert
sql = "insert into student(scode,sname,sdept,sphone,saddress) values(%s,%s,%s,%s,%s)"

curs.execute(sql,("b001","유비","국문학과","010-1111","서울"))
curs.execute(sql,("b002","관우","심리학과","010-2222","서울"))
curs.execute(sql,("b003","장비","경제학과","010-3333","서울"))
conn.commit()


sql = "select * from student"
curs.execute(sql)
rows = curs.fetchall()
print(rows)

conn.close()

# Insert 여러줄
dbPath = "../Data/test.sqlite"
conn = sqlite3.connect(dbPath)
conn = pymysql.connect(
    host = "127.0.0.1",
    user = "root",
    password = "qwer1234",
    db = "education",
    charset = "utf8"
)
data = (
    ('c001','조조','국문학과','016-111','경기'),
    ('c002','여포','국문학과','016-222','경기'),
    ('c003','손견','국문학과','016-333','경기'),
)
sql = "insert into student(scode,sname,sdept,sphone,saddress) values(%s,%s,%s,%s,%s)"

curs = conn.cursor()
curs.executemany(sql,data)
conn.commit()

sql = "select * from student"
curs.execute(sql)
rows = curs.fetchall()
print(rows)

conn.close()

---
### select내용을 df로 보기

# !pip install sqlalchemy

import pandas as pd
from sqlalchemy import create_engine

# mysql이 pymysql과 연결되어있고 아이디, 비밀번호, 아이피, 포트, 스키마 정보는 이러하다~
engine = create_engine("mysql+pymysql://root:qwer1234@127.0.0.1:3306/education")
conn = engine.connect()

data = pd.read_sql_table('student',conn)
data

conn.close()

### Dataframe을 Database로 Insert
- DaumList.csv 파일을 DB에 넣기

movies = pd.read_csv("../Data/DaumList.csv")
# 현재 df컬럼은 index, title; db컬럼은 seq, rank, title임
# 새 컬럼명 지정
column = ['rank','title']
movies.columns = column
movies.head

engine = create_engine("mysql+pymysql://root:qwer1234@127.0.0.1:3306/daum")
conn = engine.connect()

movies.to_sql(name = 'movie',con=engine,if_exists='append',index=False)

data = pd.read_sql_table('movie',conn)
data

---
### TinyDB
- NoSQL 문서형 데이터베이스 (like Firebase)
- MongoDB가 대표적이지만 MongoDB는 설치가 필요하고, TinyDB는 패키지만 설치하면 됨

# !pip install tinydb

# TinyDB를 사용하기 위한 라이브러리 불러오기
from tinydb import TinyDB, Query, where

# 데이터베이스 연결하기 (문서형이므로 file, json형식으로)
filePath = "../Data/tinydb.json"
db = TinyDB(filePath)

# 기존 테이블이 있다면 제거하기
db.drop_table("fruits")

# 테이블 생성하기
table = db.table("fruits")

# 테이블에 데이터 추가하기
table.insert({'name':'Banana','price':6000})
# 결과: {"fruits": {"1": {"name": "Banana", "price": 6000}}}

table.insert({'name':'Orange','price':12000})
table.insert({'name':'Mango','price':8400})
# 결과: "2": {"name": "Orange", "price": 12000}, "3": {"name": "Mango", "price": 8400}}}

# 모든 데이터 출력
print(table.all())
# [{'name': 'Banana', 'price': 6000}, {'name': 'Orange', 'price': 12000}, {'name': 'Mango', 'price': 8400}]

# Orange 검색
item = Query()
res = table.search(item.name=='Orange')
print(res)
# print(res['price'])
print(res[0]['price'])

# 가격이 8000원 이상인 과일명만 추출하기
item = Query()
res = table.search(item.price>=8000)
for i in range(0,2):
    print('-',res[i]['name'])

# 수정하기
table.update({'price':84000},item.name=="Mango")
res = table.search(item.name == "Mango")
print(res)

# 삭제하기 #1
table.remove(where('name')=="Mango")
# 삭제하기 #2
table.remove(item.name == "Orange")
print(table.all())
