==> Scrapy/08. WebAPI.ipynb <==
### OpenWeatherMap
http://openweathermap.org/      
기본적으로 유료이지만, 현재 날씨와 5일까지의 날씨는 무료로 사용가능
- json data

import requests
import json

apikey = "d5265640d43cf3f4ade53461a389e5fb"
# 날씨를 확인할 도시 지정하기
cities = ['Seoul,KR','Tokyo,JP','New York,US','Beijing,CN']


api = 'http://openweathermap.org/data/2.5/weather?1={city}&APPID={key}'

# 켈빈 온도를 섭씨 온도로 변환
k2c = lambda k: k -273.15

url = api.format(city=cities[0],key=apikey)
r = requests.get(url)
data = json.loads(r.text)
# print(data)
print(" + 도시 = ", data['name'])
print(" | 날씨 = ", data['weather'][0]['description'])
print(" | 최저 기온 = ", k2c(data['main']['temp_min']))

---
### Naver API로 위도 경도 가져오기

Client ID: 8ql50uk6tx   
Client Secret: X8wts2TLSDNGq843fmQEMw6uifYRhICmxlrDg0Fz

import urllib.request
import urllib

client_id = "8ql50uk6tx"
client_secret = "X8wts2TLSDNGq843fmQEMw6uifYRhICmxlrDg0Fz"

---
### 주소 위도 경도 찾기
- urllib.parse.quote("여기에 주소(지번)을 입력하면")
- 아래 process를 거쳐 위도 경도가 출력됨.

encText = urllib.parse.quote("서울시 영등포구 국회대로 608")
url = "https://naveropenapi.apigw.ntruss.com/map-geocode/v2/geocode?query=" + encText # json 결과
request = urllib.request.Request(url)
request.add_header("X-NCP-APIGW-API-KEY-ID",client_id)
request.add_header("X-NCP-APIGW-API-KEY",client_secret)
response = urllib.request.urlopen(request)
rescode = response.getcode()
if(rescode==200):
    response_body = response.read()
    entext = response_body.decode('utf-8')
else:
    print("Error Code:" + rescode)

import json    
data1 = json.loads(entext)
# print(data1)
data1

# 경도
lng = data1['addresses'][0]['x']
# 위도
lat = data1['addresses'][0]['y']

print(lng,lat)

# 주소 입력하면 경도 위도 반환하는 process

import pandas as pd

# 저장한 가장맛있는족발 지점 데이터 csv파일을 불러오기
df = pd.read_csv("../Data/Gajok.csv")
df.reset_index(drop=True)
address_list = [df.Address]

name_list = [df.Branch]



# 가족 지점 주소 입력

client_id = "8ql50uk6tx"
client_secret = "X8wts2TLSDNGq843fmQEMw6uifYRhICmxlrDg0Fz"
name_data = []
address_data = []
xy_data = []

for i in range(len(df)):
    name = name_list[0][i]
    address = address_list[0][i]
    try:
        encText = urllib.parse.quote(address)
        url = "https://naveropenapi.apigw.ntruss.com/map-geocode/v2/geocode?query=" + encText # json 결과
        request = urllib.request.Request(url)
        request.add_header("X-NCP-APIGW-API-KEY-ID",client_id)
        request.add_header("X-NCP-APIGW-API-KEY",client_secret)
        response = urllib.request.urlopen(request)
        rescode = response.getcode()
        if(rescode==200):
            response_body = response.read()
            entext = response_body.decode('utf-8')
            
            data = json.loads(entext)
            # 경도
            lng = data['addresses'][0]['x']
            # 위도
            lat = data['addresses'][0]['y']
            # 경도 위도 한 컬럼으로 join
            xy_join = [lng,lat]
            xy = ', '.join(xy_join)

            # 리스트에 추가
            name_data.append(name)
            address_data.append(address)
            xy_data.append(xy)

            # 지점명
            print('------------------------------')
            print('지점명: ',name,'\n지점주소:\t',address,'\n위도/경도: ',lng,lat)

        else:
            print("Error Code:" + rescode)
    except:
        print('---------------Error---------------')
        print(name)
        print(address)
        print('---------------Error---------------')
        pass

gajok2 = pd.DataFrame(data=zip(name_data,address_data,xy_data), columns=["Branch","Address","Location"], index=None)


gajok2.to_csv("../Data/Gajok2.csv",encoding="utf-8",index=False)

pd.read_csv("../Data/Gajok2.csv")

# 현재 df의 컬럼은 Branch, Address, Tel
# gajok2의 컬럼은 Branch, Address, Location

print(df.shape)
# df: 403rows, 3cols

print(gajok2.shape)
# gajok2: 390rows, 3cols

# merge를 통해 공통된 컬럼 (Branch, Address)를 기준으로 Dataframe 병합
gajok3 = pd.merge(df,gajok2,on=("Branch","Address"))
print(gajok3.shape)
# gajok3 결과: 390rows, 4cols
gajok3.to_csv("../Data/Gajok3.csv",encoding="utf-8",index=False)

pd.read_csv("../Data/Gajok3.csv")
